Revert "Resolve (#4)"
This reverts commit 7cae26ebcc12ca2be9ca4df345909c4528166c02.
