+++ redis_write_behind/pipelines/HashPipeline.py
@@ -1,10 +1,21 @@
-from redis_write_behind.utils.basic_utils import *
-from redis_write_behind.pipelines.PipelinesBase import *
-from redis_write_behind.pipelines.constants import *
-
 from redisgears import executeCommand as execute
+from redis_write_behind.utils.basic_utils import *
 import json
 import uuid
+
+ackExpireSeconds = 3600
+
+def SafeDeleteKey(key):
+    '''
+    Deleting a key by first renaming it so we will not trigger another execution
+    If key does not exists we will get an execution and ignore it
+    '''
+    try:
+        newKey = '__{%s}__' % key
+        execute('RENAME', key, newKey)
+        execute('DEL', newKey)
+    except Exception:
+        pass
 
 def ValidateHash(r):
     key = r['key']
@@ -48,14 +59,236 @@ def ValidateHash(r):
 
     return True
 
-class HashWriteBehind(RGWriteBase):
-    def __init__(self, GB, keysPrefix, mappings, connector, name, version=None, 
-                 primaryCacheKey=False,
-                 onFailedRetryInterval=DEFAULT_ON_FAILED_RETRY_INTERVAL, 
-                 batch=DEFAULT_BATCH, 
-                 duration=DEFAULT_DURATION_IN_MS, 
-                 transform=lambda r: r, 
-                 eventTypes=HASH_EVENT_TYPES):
+def DeleteHashIfNeeded(r):
+    key = r['key']
+    operation = r['value'][OP_KEY]
+    if operation == OPERATION_DEL_REPLICATE:
+        SafeDeleteKey(key)
+
+def ShouldProcessHash(r):
+    key = r['key']
+    value = r['value']
+    uuid = value[UUID_KEY]
+    operation = value[OP_KEY]
+    res = True
+
+    if operation == OPERATION_DEL_NOREPLICATE:
+        # we need to just delete the key but delete it directly will cause
+        # key unwanted key space notification so we need to rename it first
+        SafeDeleteKey(key)
+        res = False
+
+    if operation == OPERATION_UPDATE_NOREPLICATE:
+        res = False
+
+    if not res and uuid != '':
+        # no replication to connector is needed but ack is require
+        idToAck = '{%s}%s' % (key, uuid)
+        execute('XADD', idToAck, '*', 'status', 'done')
+        execute('EXPIRE', idToAck, ackExpireSeconds)
+
+    return res
+
+def RegistrationArrToDict(registration, depth):
+    if depth >= 2:
+        return registration
+    if type(registration) is not list:
+        return registration
+    d = {}
+    for i in range(0, len(registration), 2):
+        d[registration[i]] = RegistrationArrToDict(registration[i + 1], depth + 1)
+    return d
+
+def CompareVersions(v1, v2):
+    # None version is less then all version
+    if v1 is None:
+        return -1
+    if v2 is None:
+        return 1
+
+    if v1 == '99.99.99':
+        return 1
+    if v2 == '99.99.99':
+        return -1
+
+    v1_major, v1_minor, v1_patch = v1.split('.')
+    v2_major, v2_minor, v2_patch = v2.split('.')
+
+    if int(v1_major) > int(v2_major):
+        return 1
+    elif int(v1_major) < int(v2_major):
+        return -1
+
+    if int(v1_minor) > int(v2_minor):
+        return 1
+    elif int(v1_minor) < int(v2_minor):
+        return -1
+
+    if int(v1_patch) > int(v2_patch):
+        return 1
+    elif int(v1_patch) < int(v2_patch):
+        return -1
+
+    return 0
+
+def UnregisterOldVersions(name, version):
+    WriteBehindLog('Unregistering old versions of %s' % name)
+    registrations = execute('rg.dumpregistrations')
+    for registration in registrations:
+        registrationDict = RegistrationArrToDict(registration, 0)
+        descStr = registrationDict['desc']
+        try:
+            desc = json.loads(descStr)
+        except Exception as e:
+            continue
+        if 'name' in desc.keys() and name in desc['name']:
+            WriteBehindLog('Version auto upgrade is not atomic, make sure to use it when there is not traffic to the database (otherwise you might lose events).', logLevel='warning')
+            if 'version' not in desc.keys():
+                execute('rg.unregister', registrationDict['id'])
+                WriteBehindLog('Unregistered %s' % registrationDict['id'])
+                continue
+            v = desc['version']
+            if CompareVersions(version, v) > 0:
+                execute('rg.unregister', registrationDict['id'])
+                WriteBehindLog('Unregistered %s' % registrationDict['id'])
+            else:
+                raise Exception('Found a version which is greater or equals current version, aborting.')
+    WriteBehindLog('Unregistered old versions')
+
+def CreateAddToStreamFunction(self, primaryCacheKey):
+    def func(r):
+        data = []
+        data.append([ORIGINAL_KEY, r['key']])
+        if primaryCacheKey:
+            data.append([self.connector.PrimaryKey(), r['key']])
+        else:
+            # after key prefix value
+            data.append([self.connector.PrimaryKey(), r['key'].split(self.keysPrefix)[-1]])
+        if 'value' in r.keys():
+            value = r['value']
+            uuid = value.pop(UUID_KEY, None)
+            op = value[OP_KEY]
+            data.append([OP_KEY, op])
+            keys = value.keys()
+            if uuid is not None:
+                data.append([UUID_KEY, uuid])
+            if op == OPERATION_UPDATE_REPLICATE:
+                for kInHash, kInDB in self.mappings.items():
+                    if kInHash.startswith('_'):
+                        continue
+                    if kInHash not in keys:
+                        msg = 'AddToStream: Could not find %s in hash %s' % (kInHash, r['key'])
+                        WriteBehindLog(msg)
+                        raise Exception(msg)
+                    data.append([kInDB, value[kInHash]])
+        execute('xadd', self.GetStreamName(self.connector.TableName()), '*', *sum(data, []))
+    return func
+
+def CreateWriteDataFunction(connector, dataKey=None):
+    def func(data):
+        idsToAck = []
+        for d in data:
+            originalKey = d['value'].pop(ORIGINAL_KEY, None)
+            uuid = d['value'].pop(UUID_KEY, None)
+            if uuid is not None and uuid != '':
+                idsToAck.append('{%s}%s' % (originalKey, uuid))
+
+        # specifically, to not updating all the old WriteData calls
+        # due to JSON
+        if dataKey is None:
+            connector.WriteData(data)
+        else:
+            connector.WriteData(data, dataKey)
+
+        for idToAck in idsToAck:
+            execute('XADD', idToAck, '*', 'status', 'done')
+            execute('EXPIRE', idToAck, ackExpireSeconds)
+
+    return func
+
+class RGWriteBase():
+    def __init__(self, mappings, connector, name, version=None):
+        UnregisterOldVersions(name, version)
+
+        self.connector = connector
+        self.mappings = mappings
+
+        try:
+            self.connector.PrepereQueries(self.mappings)
+        except Exception as e:
+            # cases like mongo, that don't implement this, silence the warning
+            if "object has no attribute 'PrepereQueries'" in str(e):
+                return
+            WriteBehindLog('Skip calling PrepereQueries of connector, err="%s"' % str(e))
+
+def DeleteKeyIfNeeded(r):
+    if r['value'][OP_KEY] == OPERATION_DEL_REPLICATE:
+        # we need to just delete the key but delete it directly will cause
+        # key unwanted key space notification so we need to rename it first
+        SafeDeleteKey(r['key'])
+
+def PrepareRecord(r):
+    key = r['key']
+    value = r['value']
+
+    realKey = key.split('{')[1].split('}')[0]
+
+    realVal = execute('hgetall', realKey)
+    realVal = {realVal[i]:realVal[i + 1] for i in range(0, len(realVal), 2)}
+
+    realVal.update(value)
+
+    # delete temporary key
+    execute('del', key)
+
+    return {'key': realKey, 'value': realVal}
+
+
+
+def UpdateHash(r):
+    key = r['key']
+    value = r['value']
+    operation = value.pop(OP_KEY, None)
+    uuid_key = value.pop(UUID_KEY, None)
+    if operation == OPERATION_DEL_REPLICATE or operation == OPERATION_DEL_NOREPLICATE:
+        SafeDeleteKey(key)
+    elif operation == OPERATION_UPDATE_REPLICATE or OPERATION_UPDATE_NOREPLICATE:
+        # we need to write to temp key and then rename so we will not
+        # trigger another execution
+        tempKeyName = 'temp-{%s}' % key
+        elemets = []
+        for k,v in value.items():
+            elemets.append(k)
+            elemets.append(v)
+        execute('hset', tempKeyName, *elemets)
+        execute('rename', tempKeyName, key)
+    else:
+        msg = "Unknown operation"
+        WriteBehindLog(msg)
+        raise Exception(msg)
+
+def WriteNoReplicate(r):
+    if ShouldProcessHash(r):
+        # return true means hash should be replicate and we need to
+        # continue processing it
+        return True
+    # No need to replicate hash, just write it correctly to redis
+    operation = r['value'][OP_KEY]
+    if operation == OPERATION_UPDATE_NOREPLICATE:
+        UpdateHash(r)
+    elif operation == OPERATION_DEL_NOREPLICATE:
+        # OPERATION_DEL_NOREPLICATE was handled by ShouldProcessHash function
+        pass
+    else:
+        msg = "Unknown operation"
+        WriteBehindLog(msg)
+        raise Exception(msg)
+    return False
+
+class DefaultWriteBehind(RGWriteBase):
+    def __init__(self, GB, keysPrefix, mappings, connector, name, version=None, primaryCacheKey=False,
+                 onFailedRetryInterval=5, batch=100, duration=100, transform=lambda r: r, 
+                 eventTypes=['hset', 'hmset', 'del', 'change']):
         '''
         Register a write behind execution to redis gears
 
@@ -109,9 +342,6 @@ class HashWriteBehind(RGWriteBase):
         version - The version to set to the new created registration. Old versions with the same
                name will be removed. 99.99.99 is greater then any other version (even from itself).
 
-        primaryCacheKey - If True The key in redis will be supplied as primary key.
-                          If False part of key will be supplied as primary key.
-
         batch - the batch size on which data will be writen to target
 
         duration - interval in ms in which data will be writen to target even if batch size did not reached
@@ -122,7 +352,6 @@ class HashWriteBehind(RGWriteBase):
 
         eventTypes - The events for which to trigger
         '''
-
         UUID = str(uuid.uuid4())
         self.GetStreamName = CreateGetStreamNameCallback(UUID)
 
@@ -137,8 +366,8 @@ class HashWriteBehind(RGWriteBase):
         GB('KeysReader', desc=json.dumps(descJson)).\
         map(transform).\
         filter(ValidateHash).\
-        filter(ShouldProcessData).\
-        foreach(DeleteDataIfNeeded).\
+        filter(ShouldProcessHash).\
+        foreach(DeleteHashIfNeeded).\
         foreach(CreateAddToStreamFunction(self, primaryCacheKey)).\
         register(mode='sync', prefix='%s*' % keysPrefix, eventTypes=eventTypes, convertToStr=False)
