+++ redis_write_behind/pipelines/__init__.py
@@ -1,3 +1,2 @@
-from redis_write_behind.pipelines.HashPipeline import HashWriteBehind
+from redis_write_behind.pipelines.HashPipeline import DefaultWriteBehind
 from redis_write_behind.pipelines.KeyValuePipeline import KeyValueWriteBehind
-from redis_write_behind.pipelines.ListPipeline import ListWriteBehind
